backend/requirements.txt
alembic==1.14.0
astroid==3.3.8
backend==0.2.4.1
beautifulsoup4==4.12.3
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
coverage==7.6.10
debugpy==1.8.12
Deprecated==1.2.15
dill==0.3.9
Flask==3.1.0
Flask-Cors==5.0.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-grpc==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-wsgi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pylint==3.3.3
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
soupsieve==2.6
SQLAlchemy==2.0.36
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
